test: refactor htmlSerializer exports
